fix main screen & other articles button
