+++ lib/view_models/main_view_model.dart
@@ -2,11 +2,11 @@ import 'package:disappear/screens/home/home_screen.dart';
 import 'package:flutter/material.dart';
 
 class MainViewModel extends ChangeNotifier {
-  final List<Map<String, dynamic>> _screens = [
-    { 'title': 'Beranda', 'screen': const HomeScreen() },
-    { 'title': 'Tantangan', 'screen': const HomeScreen() },
-    { 'title': 'Pesanan', 'screen': const HomeScreen() },
-    { 'title': 'Profil', 'screen': const HomeScreen() },
+  final List<Widget> _screens = [
+    const HomeScreen(),
+    const HomeScreen(),
+    const HomeScreen(),
+    const HomeScreen(),
   ];
 
   int _selectedScreenIndex = 0;
@@ -18,7 +18,5 @@ class MainViewModel extends ChangeNotifier {
     notifyListeners();
   }
 
-  String get currentScreenTitle => _screens[_selectedScreenIndex]['title'];
-
-  Widget get currentScreen => _screens[_selectedScreenIndex]['screen'];
+  Widget get currentScreen => _screens[_selectedScreenIndex];
 }
